backend/requirements.txt.txt
fastapi==0.111.0
uvicorn==0.30.6
python-dotenv==1.0.1

# Supabase stack
supabase==2.10.0
gotrue==2.11.4
postgrest==0.18.0
realtime==2.22.0
storage3==0.9.0

# HTTP stack compat (fix do 'proxy' em gotrue)
httpx==0.28.1
httpcore==1.0.9
websockets==15.0.1
requests==2.32.3

# FastAPI deps
pydantic==2.12.0
pydantic-core==2.41.1
starlette==0.37.2
anyio==4.11.0
ujson==5.11.0
orjson==3.11.3
python-multipart==0.0.20

# OpenAI SDK
openai==1.52.2
